Long in UserRepo and axpanded column length for location

+++ src/main/kotlin/ar/edu/unsam/phm/repository/CrudUserRepository.kt
@@ -4,7 +4,7 @@ import ar.edu.unsam.phm.domain.User
 import org.springframework.data.repository.CrudRepository
 import java.util.Optional
 
-interface CrudUserRepository: CrudRepository<User, Int> {
+interface CrudUserRepository: CrudRepository<User, Long> {
     fun findByEmail(email: String): Optional<User>
 
 }
